Propagate top-level comments from proto schemas
If the source schema has leading or leading detached comments on the
protobuf syntax or package declarations, propagate them to the generated
code. This matches the behavior of protoc-gen-go and protoc-gen-go-grpc,
with one small change: we keep protobuf package comments adjacent to the
Go package declaration, so that they appear as the package-level
documentation in GoDoc.

Fixes #470.