Generate constants for procedure names
Generate constants for each RPC's fully-qualified name.

This is particularly useful in Connect interceptors and net/http
middleware, where users often want their code to behave differently
depending on the RPC being called. (For example, they may have a subset
of procedures exempted from authentication.) Without this PR, the best
approach we can offer is either hard-coding a string literal (fragile if
the package or RPC is moved, renamed, or deleted) or reflection using
the descriptor embedded in protoc-gen-go's output.

protoc-gen-go-grpc recently began doing something very similar to this,
so it seems reasonable for us to do the same.

Of the open issues, this fixes #326.