Canonicalize Connect Streaming Trailer names (#528)
After testing many approaches, I think this is the most reasonable one.
At least with Go JSON, there is no reasonable approach I could find that
will give better performance or memory allocation characteristics to do
this in one single pass. This approach should also have a _relatively_
small impact on trailers that are already canonicalized, since they will
trivially fall into the `CanonicalMIMEHeaderKey` fast-path. The impact
on the response path will probably be measured in hundreds of
nanoseconds to microseconds in most cases, which is not ideal, but we
definitely have bigger fish to fry, and correctness is important.

Closes #524.