Simplify code when buffers aren't freed
In some error handling cases, we're retrieving a value from the
sync.Pool and explicitly not returning it. Instead, we might simplify
things and just allocate a buffer (since we already know the size) and
use it and not put extra work on the pool to garbage collect objects
that are retrieved but never returned.