Generate aliases for connect.Request/Response
Reduce Connect's generics-induced wordiness by generating type aliases
for `connect.Request` and `connect.Response`.

For an actual net reduction in wordiness, we can't generate long
identifiers. That reduces our ability to manage name collisions, so we
only generate aliases for messages that are declared in the same file
and used exclusively as requests or responses (but not both). Notably,
we don't attempt to generate aliases for the stream types - they end up
even wordier than the generic types, and they end up very confusingly
named.