Include Go 1.21.0 in CI (#566)
Include Go 1.21.0 in CI, and update `golangci-lint` to a version that supports the new runtime. Along the way, clean up the Makefile a bit and update some tool dependencies.

As part of this change, we're committing to support for the last 3 versions of Go. This ends up including a version of Go that no longer receives security patches, but it matches AWS and some large companies' very slow upgrade cycles.