Set X-User-Agent in gRPC-Web clients (#572)
The gRPC-Web pseudo-specification isn't totally clear, but it sounds
like it requires setting `X-User-Agent` instead of `User-Agent` as a
part of the protocol. This makes sense in browsers, where JS _can't_ set
`User-Agent`, but it's very odd for backend clients. This PR proposes
splitting the difference and setting both headers.