Fix flaky `TestStreamForServer` test (#627)
This test was flaky due to unexpected cases where calling `Send(nil)` to
send request headers could return a non-nil error, even though the server
did not actually send back an error. If the handler immediately returned
without accepting any messages, the end-of-stream could arrive quickly
enough that the call to `Send` would notice and return `io.EOF`. But since
the call to `Send` was intended to initiate the request and the server did
not actually return any error, this is confusing to observe on the client. So
now a call to `Send(nil)` that initiates the call will return a nil error. And
any error (including EOF) already received from the server will be returned
by the next call to `Send` or `Receive`.