Wrap errors with context cancellation codes
This PR wraps errors with the appropariate connect code of Cancelled or
DeadlineExceeded if the context error is not nil.

Improves error handling for some well known error cases that do not
surface context.Cancelled errors. For example HTTP2 "client disconnect" string
errors are now raised with a Cancelled code not an Unknwon. This lets handlers
check the error code for better handling and reporting of errors.