Implement unary HTTP calls with retry support
Changes the internal duplexHTTPCall to switch on non streaming client
requests to block and wait for the response. This avoids the need to
convert the reader to a writer with io.Pipe and the go routine to run
asynchronously. On unary requests we are now able to set the
`Content-Length` header and the `GetBody` function for retries.

To safely reuse the payload buffer a new type `payloadCloser` is added
to implement the required HTTP body semantics. On receiving a response
the request may still be read up and until the response is body is
closed. To ensure the request body is safe to releasee we wait for a
complete read or close on the request body. Retries may read, close or
rewind the body multiple times before a response is returned.