Fix ErrorWriter connect GET protocol classifier
ErrorWriter will now correctly classifying connect GET requests if
the connect protocol version header is present or the connect protocol
version is set in the URL query parameters. To handle connect GET
requests without the protocol version the ErrorWriter will write errors
as connect unary errors if the protocol is unknown. This ensures the
error is written to the client. The connect unary error payload is
always in JSON and therefore human-readable to the caller. Clients are
still recommended to check `IsSupported` first to ensure the protocol
will match the desired and to provide fallback to other error writers.