Implement unary HTTP calls with retry support (#649)
Internal changes to the HTTP client to remove some of the overhead
required for streaming calls but unnecessary for non-streaming client
requests. This PR changes both of the non-streaming client request
types: unary and server streams. These types will now create the client
request with the message payload upfront before calling `client.Do`.
This removes the overhead of the `io.Pipe` to convert the body from a
reader to a writer and the extra goroutine required to write to the pipe
asynchronously. From this we get improved functionality for
non-streaming requests by setting the `Content-Length` header and
implementing `GetBody` function to enable retries.