Wrap errors with context cancellation codes (#659)
This PR wraps errors with the appropriate connect code of Cancelled or
DeadlineExceeded if the context error is not nil.

Improves error handling for some well known error cases that do not
surface context.Cancelled errors. For example HTTP2 "client disconnect"
string errors are now raised with a Cancelled code not an Unknown. This
lets handlers check the error code for better handling and reporting of
errors.

Fix for #645