In Connect unary protocol, fallback to code based on HTTP status if unable to deserialize it (#702)
The spec states the following:

When reading data from the wire, client implementations must
use the HTTP-to-Connect mapping to infer a Connect error
code if Bare-Message is missing or malformed.

However, connect-go was only considering the message to be
malformed if json.Unmarshal failed. Because Go's encoding/json
package is fairly lenient in many ways, there were many kinds of
malformed responses that would be accepted, such as the
"code" property being absent or null. In these cases, connect-go
was always falling back to using "unknown" as the code, instead
of using the HTTP-to-Connect mapping as required by the spec.
This addresses that and reconciles connect-go's behavior with
the specification.