Update a couple of more places where we should return a context error if there is one (#709)
With an HTTP/2 server implementation that eagerly performs a
server-side cancelation of the stream upon seeing its deadline
elapse, the client could end up returning a "canceled" error code,
due to receiving a "RSTStream" frame, but should ideally return
a "deadline exceeded" error instead. This fixes things so that it
will instead return "deadline exceeded" in this situation.