Add limited support for Bidi streams over HTTP1
This adds limited support for bidirectional streams over HTTP1
transports. It removes the explicit error response which disabled this
behavior based on the protocol stream type. Bidi streams may now be used
in half-duplex mode. This does not enable full-duplex streaming. All
requests must be sent before responses are received.

This removes the workarounds in the conformance test suite to fake a
HTTP2 request/response in order to allow for this behavior. In doing so
a deadlock on writing to a closed requests was fixed.

Signed-off-by: Edward McFarlane <emcfarlane@buf.build>