Add package_suffix option (#804)
This is the alternative option `package_suffix` re:
https://github.com/connectrpc/connect-go/pull/803#pullrequestreview-2512877564.
I'm not sure which one I like more - I don't like the name
`same_package` that much, and this gives more flexibility, but I don't
know if we want to offer that flexibility. Looking for opinions.

Regardless of if we choose this or #803 as-is, we should make sure the
README.md documents the option.

Signed-off-by: bufdev <bufdev-github@buf.build>
Signed-off-by: Edward McFarlane <emcfarlane@buf.build>