Add same_package option to protoc-gen-connect-go (#803)
This adds an option `package_suffix` to `protoc-gen-connect-go`. It allows to
customize the output directory for the generated code. The sub-package of the
package containing the base .pb.go files using the given suffix. An empty
suffix denotes to generate into the same package as the base pb.go files.
Default is "connect".

---------

Signed-off-by: bufdev <bufdev-github@buf.build>
Signed-off-by: Edward McFarlane <emcfarlane@buf.build>
Co-authored-by: Edward McFarlane <emcfarlane@buf.build>