We don't support 1.20 anymore, so stop testing with it in Windows (#818)
We just missed this when we updated to Go 1.21. It was passing anyway
because releases prior to 1.21 are lenient -- in that they will try to
build and run code even if the `go` directive in `go.mod` indicates a
newer version. And since none of the code was actually relying on new
APIs, features, or semantics in Go 1.21, the tests would pass.

This also includes a change that `go mod tidy` keeps wanting to make to
`internal/conformance/go.mod`. Our "checkgenerate" make target was
previously catching these in CI only in the main module. This PR includes
a small change to `Makefile` so that this kind of thing can be caught in
the conformance module going forward, too.