Fix issues with 'make generate' (#819)
This fixes issues in `make generate` that led to some generated
code in here being incorrect (generated from incorrect version of
protoc-gen-go) yet it was not being caught in CI.

The issues was that plugins were not made available in PATH to
the steps in a separate generate.sh script. And then that script
was failing to return a non-zero exit code when some of the steps
therein failed.