Standardize on For over From (#867)
This is in relation to
https://github.com/connectrpc/connect-go/pull/851. This standardizes on
using `For` instead of `From` for all of our functions. The two words
are interchangeable, and all of our public methods use `For`, so this
has all of our private methods do so as well. A new public method
`CallInfoFromHandler` is/was being introduced in #851, and in relation
to renaming it to `CallInfoForHandler`, this furthers the
standardization.

Signed-off-by: bufdev <bufdev-github@buf.build>
Signed-off-by: Steve Ayers <sayers@buf.build>